Switch from single-precision to double-precision floating point numbers

+++ src/main/java/org/marble/entity/Entity.java
@@ -1,22 +1,22 @@
 package org.marble.entity;
 
-import javax.vecmath.Matrix4f;
+import javax.vecmath.Matrix4d;
 
 import org.marble.Game;
 
 /**
  * An entity located somewhere in the simulated world.
- *
+ * 
  * A class implementing this interface must have one or more public constructors
  * that take zero or more arguments of the types
- * {@code [String, Vector3f, Float]} that will be called when the entity is
+ * {@code [String, Vector3d, double]} that will be called when the entity is
  * loaded.
  */
 public interface Entity {
     /**
      * The current location and rotation of the entity.
      */
-    public Matrix4f getTransform();
+    public Matrix4d getTransform();
 
     /**
      * Performs allocation of resources. This method is called when the entity
@@ -27,5 +27,5 @@ public interface Entity {
     /**
      * Sets the location and rotation of the entity.
      */
-    public void setTransform(Matrix4f transform);
+    public void setTransform(Matrix4d transform);
 }
