+++ src/main/java/org/marble/util/Direction.java
@@ -1,6 +1,6 @@
 package org.marble.util;
 
-import javax.vecmath.Vector3f;
+import javax.vecmath.Vector3d;
 
 /**
  * Represents a cardinal direction.
@@ -8,20 +8,20 @@ import javax.vecmath.Vector3f;
 public enum Direction {
     NORTH(0, 1, 0), EAST(1, 0, 0), SOUTH(0, -1, 0), WEST(-1, 0, 0),
     UP(0, 0, 1), DOWN(0, 0, -1);
-    private Vector3f physicalDirection;
+    private Vector3d physicalDirection;
 
-    private Direction(final float x, final float y, final float z) {
-        physicalDirection = new Vector3f(x, y, z);
+    private Direction(final double x, final double y, final double z) {
+        physicalDirection = new Vector3d(x, y, z);
     }
 
     /**
      * A normalized vector representing the physical direction of this cardinal
      * direction.
-     *
+     * 
      * @return A mutable reference to the physical direction. This vector must
      *         not be altered.
      */
-    public Vector3f getPhysicalDirection() {
+    public Vector3d getPhysicalDirection() {
         return physicalDirection;
     }
 }
